requirements.txt
requests>=2.28.0
pyyaml>=6.0
pillow>=9.0.0
